* Made some progress on the Telnet listener. Now special telnet chars ( > 0x7F) are handled properly and won't be in the ASCII string. Just a few more tweaks to the way some ASCII chars are handled and it will be ready...  * Fixed NullPointerException in SkillDataLoader.  * Fixed MCDB loading of mobs and skills.  * Added classes for inventory (items, equips, pets, rings, etc.) and moved Player class from argonms.game to argonms.character.  * Support for KvjCompiler r14. Changes include Character.wz compiling and reduced size of Mob files.  * Character.wz is now processed in KvjItemDataLoader. Taming Mob IDs are now read in McdbItemDataLoader as well. In McdbItemDataLoader, equips can now be made tradeBlocked.

+++ src/argonms/loading/mob/Skill.java
@@ -23,22 +23,22 @@ package argonms.loading.mob;
  * @author GoldenKevin
  */
 public class Skill {
-	private int skill;
-	private int level;
+	private short skill;
+	private byte level;
 
-	public void setSkill(int id) {
+	public void setSkill(short id) {
 		this.skill = id;
 	}
 
-	public void setLevel(int level) {
+	public void setLevel(byte level) {
 		this.level = level;
 	}
 
-	public int getSkill() {
+	public short getSkill() {
 		return skill;
 	}
 
-	public int getLevel() {
+	public byte getLevel() {
 		return level;
 	}
